check tempfiles before removal and fix another 'invalid char' bug in the http arguments

+++ crispor.py
@@ -2707,7 +2707,8 @@ def findOfftargetsBwa(queue, batchId, batchBase, faFname, genome, pam, bedFname)
     # remove the temporary files
     tempFnames = [saFname, matchesBedFname, filtMatchesBedFname]
     for tfn in tempFnames:
-        os.remove(tfn)
+        if isfile(tfn):
+            os.remove(tfn)
     return bedFname
 
 def makeVariants(seq):
@@ -3502,10 +3503,11 @@ def findVariantsInRange(vcfFname, chrom, start, end, strand, minFreq):
         errAbort("%s not found" % vcfFname)
     tb = tabix.open(vcfFname)
     chrom = chrom.replace("chr","")
-    #try:
-    records = tb.query(chrom, start+1, end) # VCF is 1-based
-    #except tabix.TabixError:
-        #records = []
+    try:
+        records = tb.query(chrom, start+1, end) # VCF is 1-based
+    except tabix.TabixError:
+        sys.stderr.write("Chromosome in query does not exist in VCF file? chrom: %s, VCF file: %s\n" % (chrom, vcfFname))
+        records = []
         
 
     varDict = defaultdict(list)
@@ -4629,7 +4631,7 @@ def cgiGetSelfUrl(changeParams, anchor=None, onlyParams=None):
         newParams = mergeParamDicts(cgiSubs, changeParams)
     else:
         newParams = mergeParamDicts(cgiParams, changeParams)
-    paramStrs = ["%s=%s" % (key, val) for key, val in newParams.iteritems()]
+    paramStrs = ["%s=%s" % (key, urllib.quote(val)) for key, val in newParams.iteritems()]
     paramStr = "&".join(paramStrs)
     url = basename(__file__)+"?"+paramStr
     if anchor is not None:
